requirements.txt
Flask>=2.1
Pillow>=9.0
